Move sensor display in dedicated package
