+++ Amio/app/src/main/res/layout/fragment_sensor_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".ui.main.SensorFragment"
+    tools:context=".ui.main.sensor.SensorFragment"
     tools:listitem="@layout/sensor_view_holder" />
